Consistently Capitalizing Component Constantly
