+++ lib/middleware-linkage/input/interactive/components/Label.js
@@ -3,7 +3,7 @@
 
 import etch from 'etch'
 import { Label  as _Label} from '@purduesigbots/pros-cli-middleware'
-import { Component } from './component'
+import { Component } from './Component'
 
 
 export class Label extends _Label {
