+++ lib/middleware-linkage/input/interactive/components/ParameterizedComponent.js
@@ -4,7 +4,7 @@
 import etch from 'etch'
 import TextEditorWrapper from './TextEditorWrapper'
 import { ParameterizedComponent  as _ParameterizedComponent} from '@purduesigbots/pros-cli-middleware'
-import { Component } from './component'
+import { Component } from './Component'
 
 export class ParameterizedComponent extends _ParameterizedComponent {
   constructor(args) {
