+++ lib/middleware-linkage/input/interactive/components/index.js
@@ -1,6 +1,6 @@
 'use babel'
 
-import { Component } from './component'
+import { Component } from './Component'
 import { ParameterizedComponent } from './ParameterizedComponent'
 import { BasicParameterizedComponent } from './BasicParameterizedComponent'
 import { InputBox } from './InputBox'
